requirements.txt
asgiref==3.2.10
coverage==5.3
Django==3.1.1
django-debug-toolbar==3.1.1
django-defender==0.7.0
getsecret==0.0.6
ldap3==2.8.1
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
PyYAML==5.3.1
q==2.6
redis==3.5.3
six==1.15.0
sqlparse==0.3.1
